R4R: convert uint64 to int64 for BSON no support uint64

+++ modules/oracle/create_feed.go
@@ -7,7 +7,7 @@ import (
 
 type DocMsgCreateFeed struct {
 	FeedName          string       `bson:"feed_name" yaml:"feed_name"`
-	LatestHistory     uint64       `bson:"latest_history" yaml:"latest_history"`
+	LatestHistory     int64        `bson:"latest_history" yaml:"latest_history"`
 	Description       string       `bson:"description"`
 	Creator           string       `bson:"creator"`
 	ServiceName       string       `bson:"service_name" yaml:"service_name"`
@@ -15,7 +15,7 @@ type DocMsgCreateFeed struct {
 	Input             string       `bson:"input"`
 	Timeout           int64        `bson:"timeout"`
 	ServiceFeeCap     []types.Coin `bson:"service_fee_cap" yaml:"service_fee_cap"`
-	RepeatedFrequency uint64       `bson:"repeated_frequency" yaml:"repeated_frequency"`
+	RepeatedFrequency int64        `bson:"repeated_frequency" yaml:"repeated_frequency"`
 	AggregateFunc     string       `bson:"aggregate_func" yaml:"aggregate_func"`
 	ValueJsonPath     string       `bson:"value_json_path" yaml:"value_json_path"`
 	ResponseThreshold uint32       `bson:"response_threshold" yaml:"response_threshold"`
@@ -30,7 +30,7 @@ func (m *DocMsgCreateFeed) BuildMsg(v interface{}) {
 	msg := v.(*MsgCreateFeed)
 
 	m.FeedName = msg.FeedName
-	m.LatestHistory = msg.LatestHistory
+	m.LatestHistory = int64(msg.LatestHistory)
 	m.Description = msg.Description
 	m.Creator = msg.Creator
 	m.ServiceName = msg.ServiceName
@@ -38,7 +38,7 @@ func (m *DocMsgCreateFeed) BuildMsg(v interface{}) {
 	m.Input = msg.Input
 	m.Timeout = msg.Timeout
 	m.ServiceFeeCap = types.BuildDocCoins(msg.ServiceFeeCap)
-	m.RepeatedFrequency = msg.RepeatedFrequency
+	m.RepeatedFrequency = int64(msg.RepeatedFrequency)
 	m.AggregateFunc = msg.AggregateFunc
 	m.ValueJsonPath = msg.ValueJsonPath
 	m.ResponseThreshold = msg.ResponseThreshold
